% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_w1.R
\docType{data}
\name{content_w1}
\alias{content_w1}
\title{Content of matchday 1}
\format{
A nested list
\describe{
\item{matches}{List of game details each with length 7}
...
}
}
\source{
\url{https://rapidapi.com/heisenbug/api/Premier\%20League\%20Live\%20Scores}
}
\usage{
content_w1
}
\description{
Content of a GET request for the Premier League Live Scores API for matchday 1
}
\keyword{datasets}
