% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_w25.R
\docType{data}
\name{content_w25}
\alias{content_w25}
\title{Content of matchday 25}
\format{
An object of class \code{list} of length 1.
}
\usage{
content_w25
}
\description{
See documentation for content_w1
}
\keyword{datasets}
