% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_results.R
\name{get_away_scores}
\alias{get_away_scores}
\title{Title}
\usage{
get_away_scores(content)
}
\arguments{
\item{content}{the value from a httr::content() call to a httr::GET() response.}
}
\description{
Description
}
