% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_results.R
\name{get_home_teams}
\alias{get_home_teams}
\title{Retrieve home teams for given matchday}
\usage{
get_home_teams(content)
}
\arguments{
\item{content}{the value from a httr::content() call to a httr::GET() response.}
}
\value{
a character vector of team names.
}
\description{
Retrieve home teams for given matchday
}
