% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_results.R
\name{get_match_kickoff}
\alias{get_match_kickoff}
\title{Title}
\usage{
get_match_kickoff(content, as_datetime = FALSE)
}
\arguments{
\item{content}{the value from a httr::content() call to a httr::GET() response.}

\item{as_datetime}{DO NOT USE YET.}
}
\description{
Description
}
