% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_results.R
\name{get_match_request}
\alias{get_match_request}
\title{Get request for given match day}
\usage{
get_match_request(matchday, key)
}
\arguments{
\item{matchday}{a string with the mathday "1"-"38"}

\item{key}{a string with your rapid api key}
}
\value{
a httr::GET() response.
}
\description{
Get request for given match day
}
